requirements.txt
# Database
psycopg2-binary==2.9.9

# Environment management
python-dotenv==1.0.0

# Optional: For database migrations and management
sqlalchemy==2.0.23

# Development
pytest==7.4.3